Minigames server side
main.py: prepared the handler. uses /template/minigames.html template,
and renders into it definitions_list which is a list of 60 definitions.
each element in the list is a pair of a definition and an array of
possible answers

minigamesUtils.py: mainly a function that generates the list above

minigames.html: usage example and poc